requirements.txt
paho-mqtt>=1.1
requests